optimized to much inits in portaudio

+++ engine/audio/src/audiocapture_portaudio.h
@@ -57,7 +57,7 @@ protected:
     /** @reimpl */
     bool readAudio(int maxSize);
 private:
-    float Volume;
+    qreal Volume;
 };
 
 /** @} */
